+++ web/src/main/resources/static/js/index.js
@@ -33793,10 +33793,8 @@
 	        return (React.createElement("div", { id: "toolbar" },
 	            React.createElement(ToolbarButton, { label: "Undo", icon: "fa-rotate-left" }),
 	            React.createElement(ToolbarButton, { label: "Redo", icon: "fa-rotate-right" }),
-	            React.createElement(ToolbarButton, { label: "Select", icon: "fa-arrow-pointer", name: "select", currentTool: this.props.tool, setTool: this.props.setTool }),
-	            React.createElement(ToolbarButton, { label: "Move", icon: "fa-up-down-left-right", name: "move", currentTool: this.props.tool, setTool: this.props.setTool }),
-	            React.createElement(ToolbarButton, { label: "New Contour", icon: "fa-plus", name: "create", currentTool: this.props.tool, setTool: this.props.setTool }),
-	            React.createElement(ToolbarButton, { label: "Show Labels", icon: "fa-eye" })));
+	            React.createElement(ToolbarButton, { label: "Edit", icon: "fa-up-down-left-right", name: "edit", currentTool: this.props.tool, setTool: this.props.setTool }),
+	            React.createElement(ToolbarButton, { label: "New Contour", icon: "fa-plus", name: "create", currentTool: this.props.tool, setTool: this.props.setTool })));
 	    };
 	    return Toolbar;
 	}(React.Component));
@@ -33847,8 +33845,8 @@
 	    CreateToolSidebar.prototype.render = function () {
 	        return (React.createElement("div", { className: "tool-sidebar tool-sidebar__create " + (this.props.isActive ? 'tool-sidebar__active' : '') },
 	            React.createElement("h3", null, "Create Tool"),
-	            React.createElement("p", null, "click on canvas to create vertexes of new contour"),
-	            React.createElement("button", null, "Close Contour")));
+	            React.createElement("p", null, "Click on canvas to create vertexes of new contour."),
+	            React.createElement("p", null, "Click on first vertex to finish creation.")));
 	    };
 	    return CreateToolSidebar;
 	}(BaseToolSidebar));
@@ -33867,19 +33865,18 @@
 	        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
 	    };
 	})();
-	var MoveToolSidebar = /** @class */ (function (_super) {
-	    __extends$6(MoveToolSidebar, _super);
-	    function MoveToolSidebar() {
+	var EditToolSidebar = /** @class */ (function (_super) {
+	    __extends$6(EditToolSidebar, _super);
+	    function EditToolSidebar() {
 	        return _super !== null && _super.apply(this, arguments) || this;
 	    }
-	    MoveToolSidebar.prototype.render = function () {
+	    EditToolSidebar.prototype.render = function () {
 	        return (React.createElement("div", { className: "tool-sidebar tool-sidebar__create " + (this.props.isActive ? 'tool-sidebar__active' : '') },
-	            React.createElement("h3", null, "Move Tool"),
-	            React.createElement("p", null, "Drag and drop contour")));
+	            React.createElement("h3", null, "Edit Tool"),
+	            React.createElement("p", null, "Drag and drop contours and vertexes of selected contour")));
 	    };
-	    return MoveToolSidebar;
+	    return EditToolSidebar;
 	}(BaseToolSidebar));
-	// }
 
 	var __extends$5 = (undefined && undefined.__extends) || (function () {
 	    var extendStatics = function (d, b) {
@@ -33929,6 +33926,7 @@
 	        return _this;
 	    }
 	    ContourList.prototype.click = function (index) {
+	        this.props.setTool("edit");
 	        this.props.selectContour(index);
 	    };
 	    ContourList.prototype.render = function () {
@@ -33967,8 +33965,8 @@
 	        return (React.createElement("div", { id: "sidebar" },
 	            React.createElement(CreateToolSidebar, { isActive: this.props.tool == "create" }),
 	            React.createElement(SelectToolSidebar, { isActive: this.props.tool == "select" }),
-	            React.createElement(MoveToolSidebar, { isActive: this.props.tool == "move" }),
-	            React.createElement(ContourList, { contours: this.props.contours, selectContour: this.props.selectContour, selectedContour: this.props.selectedContour })));
+	            React.createElement(EditToolSidebar, { isActive: this.props.tool == "edit" }),
+	            React.createElement(ContourList, { contours: this.props.contours, selectContour: this.props.selectContour, selectedContour: this.props.selectedContour, setTool: this.props.setTool })));
 	    };
 	    return Sidebar;
 	}(React.Component));
@@ -66754,56 +66752,112 @@ For more info see: https://github.com/konvajs/react-konva/issues/194
 	        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
 	    };
 	})();
+	var __spreadArrays$1 = (undefined && undefined.__spreadArrays) || function () {
+	    for (var s = 0, i = 0, il = arguments.length; i < il; i++) s += arguments[i].length;
+	    for (var r = Array(s), k = 0, i = 0; i < il; i++)
+	        for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++)
+	            r[k] = a[j];
+	    return r;
+	};
+	/// TODO class too long
 	var CanvasWrapper = /** @class */ (function (_super) {
 	    __extends$2(CanvasWrapper, _super);
 	    function CanvasWrapper(props) {
 	        var _this = _super.call(this, props) || this;
 	        _this.tempVertexes = [];
-	        _this.handleDragMoveVertex = _this.handleDragMoveVertex.bind(_this);
-	        _this.handleDragMoveContour = _this.handleDragMoveContour.bind(_this);
-	        _this.handleDragStartContour = _this.handleDragStartContour.bind(_this);
-	        _this.handleDragEndContour = _this.handleDragEndContour.bind(_this);
-	        _this.handleClickContour = _this.handleClickContour.bind(_this);
+	        _this.diamondShape = [
+	            0, -5,
+	            5, 0,
+	            0, 5,
+	            -5, 0
+	        ];
+	        _this.state = {
+	            points: [],
+	            mousePos: { x: 0, y: 0 }
+	        };
+	        _this.vertexDragMove = _this.vertexDragMove.bind(_this);
+	        _this.contourDragMove = _this.contourDragMove.bind(_this);
+	        _this.contourDragEnd = _this.contourDragEnd.bind(_this);
+	        _this.contourMouseDown = _this.contourMouseDown.bind(_this);
+	        _this.stageMouseDown = _this.stageMouseDown.bind(_this);
+	        _this.stageMouseMove = _this.stageMouseMove.bind(_this);
+	        _this.createdVertexMouseDown = _this.createdVertexMouseDown.bind(_this);
 	        return _this;
 	    }
-	    CanvasWrapper.prototype.handleDragMoveVertex = function (event, vertexIndex) {
+	    CanvasWrapper.prototype.componentDidUpdate = function (prevProps) {
+	        if (this.props.state.tool !== prevProps.state.tool) {
+	            this.setState({ points: [] });
+	        }
+	    };
+	    CanvasWrapper.prototype.vertexDragMove = function (event, vertexIndex) {
+	        if (this.props.state.tool !== "edit") {
+	            return;
+	        }
 	        this.props.setVertexPos(vertexIndex, event.target.attrs.x, event.target.attrs.y);
 	    };
-	    CanvasWrapper.prototype.handleDragMoveContour = function (event, contourIndex) {
+	    CanvasWrapper.prototype.contourDragMove = function (event, contourIndex) {
+	        if (this.props.state.tool !== "edit") {
+	            return;
+	        }
 	        this.props.setContourDragPos(contourIndex, event.target.attrs.x, event.target.attrs.y);
-	        //         console.log("move contour", [event.target.attrs.x,  event.target.attrs.y]);
-	        //         let x = event.target.attrs.x;
-	        //         let y = event.target.attrs.y;
-	        //         event.target.attrs.x = 0;
-	        //         event.target.attrs.y = 0;
-	        //         console.log("move contour 2", [event.target.attrs.x,  event.target.attrs.y]);
-	        // //         let stage = event.target.getStage();
-	        // //         stage.getPointerPosition().x;
-	        // //         this.props.setContourPos( contourIndex, x - this.prevX, y - this.prevY);
-	        //         this.prevX = x;
-	        //         this.prevY = y;
-	        // kind of works but blinking and does not match cursor location
-	        // need better solution
-	        // problem is we are moving contour zero point together with vertexes
-	        // <Line x= > does not remain zero
-	        // Maybe
-	        // Maybe we can add Line.x/y to data model temporary and "collapse" contour back to normal on drag end?
 	    };
-	    CanvasWrapper.prototype.handleDragStartContour = function (event) {
-	        this.prevX = 0;
-	        this.prevY = 0;
-	    };
-	    CanvasWrapper.prototype.handleDragEndContour = function (event, contourIndex) {
+	    CanvasWrapper.prototype.contourDragEnd = function (event, contourIndex) {
+	        if (this.props.state.tool !== "edit") {
+	            return;
+	        }
 	        this.props.collapseContour(contourIndex);
 	    };
-	    CanvasWrapper.prototype.handleClickContour = function (event, contourIndex) {
-	        console.log(contourIndex);
+	    CanvasWrapper.prototype.contourMouseDown = function (event, contourIndex) {
+	        if (this.props.state.tool !== "edit") {
+	            return;
+	        }
 	        this.props.selectContour(contourIndex);
 	    };
+	    CanvasWrapper.prototype.stageMouseDown = function (event) {
+	        if (this.props.state.tool !== "create") {
+	            return;
+	        }
+	        console.log("stageMouseDown", event);
+	        var pos = event.target.getStage().getPointerPosition();
+	        this.setState({
+	            points: __spreadArrays$1(this.state.points, [{ x: pos.x, y: pos.y }]),
+	            mousePos: { x: pos.x, y: pos.y }
+	        });
+	    };
+	    CanvasWrapper.prototype.createdVertexMouseDown = function (event, vertexIndex) {
+	        if (this.props.state.tool !== "create"
+	            || vertexIndex !== 0) {
+	            return;
+	        }
+	        console.log("first vertex clicked");
+	        event.cancelBubble = true;
+	        this.props.addContour(this.state.points);
+	        this.setState({ points: [] });
+	    };
+	    CanvasWrapper.prototype.stageMouseMove = function (event) {
+	        if (this.props.state.tool !== "create"
+	            || this.state.points.length < 1) {
+	            return;
+	        }
+	        var pos = event.target.getStage().getPointerPosition();
+	        this.setState({
+	            mousePos: { x: pos.x, y: pos.y }
+	        });
+	    };
+	    /// TODO method too long
 	    CanvasWrapper.prototype.render = function () {
 	        var _this = this;
+	        var createdPoints = [];
+	        this.state.points.forEach(function (vertex) {
+	            createdPoints.push(vertex.x);
+	            createdPoints.push(vertex.y);
+	        });
+	        if (this.state.points.length > 0) {
+	            createdPoints.push(this.state.mousePos.x);
+	            createdPoints.push(this.state.mousePos.y);
+	        }
 	        return (React.createElement("div", { id: "canvas-wrapper" },
-	            React.createElement(Stage, { width: window.innerWidth, height: window.innerHeight },
+	            React.createElement(Stage, { width: window.innerWidth, height: window.innerHeight, onMouseDown: this.stageMouseDown, onMouseMove: this.stageMouseMove },
 	                React.createElement(Layer, null,
 	                    this.props.state.contours.map(function (contour, contourIndex) {
 	                        var points = [];
@@ -66811,17 +66865,15 @@ For more info see: https://github.com/konvajs/react-konva/issues/194
 	                            points.push(vertex.x);
 	                            points.push(vertex.y);
 	                        });
-	                        return (React.createElement(Line, { key: "line" + contourIndex, points: points, stroke: "black", strokeWidth: 2, closed: true, onDragStart: _this.handleDragStartContour, onDragMove: function (e) { return _this.handleDragMoveContour(e, contourIndex); }, onDragEnd: function (e) { return _this.handleDragEndContour(e, contourIndex); }, onMouseDown: function (e) { return _this.handleClickContour(e, contourIndex); }, x: contour.x, y: contour.y, draggable: true }));
+	                        return (React.createElement(Line, { key: "line" + contourIndex, points: points, stroke: "black", strokeWidth: 2, closed: true, onDragMove: function (e) { return _this.contourDragMove(e, contourIndex); }, onDragEnd: function (e) { return _this.contourDragEnd(e, contourIndex); }, onMouseDown: function (e) { return _this.contourMouseDown(e, contourIndex); }, x: contour.x, y: contour.y, draggable: _this.props.state.tool === "edit" }));
 	                    }),
 	                    this.props.state.contours.map(function (contour, contourIndex) {
 	                        return (React.createElement(Group, { key: contourIndex }, contourIndex === _this.props.state.selectedContour ?
-	                            contour.vertexes.map(function (vertex, vertexIndex) { return (React.createElement(Line, { key: vertexIndex, points: [
-	                                    0, -5,
-	                                    5, 0,
-	                                    0, 5,
-	                                    -5, 0
-	                                ], x: vertex.x + contour.x, y: vertex.y + contour.y, fill: "#009999", stroke: "black", strokeWidth: 1, closed: true, onDragMove: function (e) { return _this.handleDragMoveVertex(e, vertexIndex); }, draggable: true })); }) : ""));
-	                    })))));
+	                            contour.vertexes.map(function (vertex, vertexIndex) { return (React.createElement(Line, { key: vertexIndex, points: _this.diamondShape, x: vertex.x + contour.x, y: vertex.y + contour.y, fill: "#009999", stroke: "black", strokeWidth: 1, closed: true, onDragMove: function (e) { return _this.vertexDragMove(e, vertexIndex); }, draggable: _this.props.state.tool === "edit" })); }) : ""));
+	                    })),
+	                this.props.state.tool === "create" ? (React.createElement(Layer, null,
+	                    React.createElement(Line, { points: createdPoints, stroke: "blue", strokeWidth: 2, closed: false }),
+	                    React.createElement(Group, { key: "createdPoints" }, this.state.points.map(function (vertex, vertexIndex) { return (React.createElement(Line, { key: vertexIndex, points: _this.diamondShape, x: vertex.x, y: vertex.y, fill: "#009999", stroke: "blue", strokeWidth: 1, closed: true, onMouseDown: function (e) { return _this.createdVertexMouseDown(e, vertexIndex); } })); })))) : "")));
 	    };
 	    return CanvasWrapper;
 	}(React.Component));
@@ -66915,6 +66967,13 @@ For more info see: https://github.com/konvajs/react-konva/issues/194
 	        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
 	    };
 	})();
+	var __spreadArrays = (undefined && undefined.__spreadArrays) || function () {
+	    for (var s = 0, i = 0, il = arguments.length; i < il; i++) s += arguments[i].length;
+	    for (var r = Array(s), k = 0, i = 0; i < il; i++)
+	        for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++)
+	            r[k] = a[j];
+	    return r;
+	};
 	var App = /** @class */ (function (_super) {
 	    __extends(App, _super);
 	    function App(props) {
@@ -66930,6 +66989,7 @@ For more info see: https://github.com/konvajs/react-konva/issues/194
 	        _this.setVertexPos = _this.setVertexPos.bind(_this);
 	        _this.setContourDragPos = _this.setContourDragPos.bind(_this);
 	        _this.collapseContour = _this.collapseContour.bind(_this);
+	        _this.addContour = _this.addContour.bind(_this);
 	        return _this;
 	    }
 	    App.prototype.setTool = function (tool) {
@@ -66958,6 +67018,18 @@ For more info see: https://github.com/konvajs/react-konva/issues/194
 	                selectedContour: prevState.selectedContour
 	            };
 	        });
+	    };
+	    App.prototype.addContour = function (vertexes) {
+	        this.setState({
+	            contours: __spreadArrays(this.state.contours, [{
+	                    title: "New Contour " + this.state.contours.length,
+	                    x: 0,
+	                    y: 0,
+	                    vertexes: vertexes
+	                }])
+	        });
+	        this.setTool("edit");
+	        this.selectContour(this.state.contours.length);
 	    };
 	    /**
 	    select contour and make it active
@@ -67005,8 +67077,8 @@ For more info see: https://github.com/konvajs/react-konva/issues/194
 	        return (React.createElement("div", { id: "main-wrapper" },
 	            React.createElement(MainMenu, { setContours: this.setContours }),
 	            React.createElement(Toolbar, { tool: this.state.tool, setTool: this.setTool }),
-	            React.createElement(Sidebar, { tool: this.state.tool, contours: this.state.contours, selectContour: this.selectContour, selectedContour: this.state.selectedContour }),
-	            React.createElement(CanvasWrapper, { state: this.state, setContours: this.setContours, setVertexPos: this.setVertexPos, setContourDragPos: this.setContourDragPos, collapseContour: this.collapseContour, selectContour: this.selectContour })));
+	            React.createElement(Sidebar, { tool: this.state.tool, contours: this.state.contours, selectContour: this.selectContour, selectedContour: this.state.selectedContour, setTool: this.setTool }),
+	            React.createElement(CanvasWrapper, { state: this.state, setContours: this.setContours, setVertexPos: this.setVertexPos, setContourDragPos: this.setContourDragPos, collapseContour: this.collapseContour, selectContour: this.selectContour, addContour: this.addContour })));
 	    };
 	    return App;
 	}(React.Component));