+++ web/src/main/ts/CreateToolSidebar.tsx
@@ -8,9 +8,11 @@ import { BaseToolSidebar } from "./BaseToolSidebar";
                 <div className={"tool-sidebar tool-sidebar__create " + ( this.props.isActive ? 'tool-sidebar__active' : '')}>
                     <h3>Create Tool</h3>
                     <p>
-                    click on canvas to create vertexes of new contour
+                    Click on canvas to create vertexes of new contour.
                     </p>
-                    <button>Close Contour</button>
+                    <p>
+                    Click on first vertex to finish creation.
+                    </p>
                 </div>
             );
         }