+++ web/src/main/ts/Sidebar.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { CreateToolSidebar } from "./CreateToolSidebar";
-import { MoveToolSidebar } from "./MoveToolSidebar";
+import { EditToolSidebar } from "./EditToolSidebar";
 import { SelectToolSidebar } from "./SelectToolSidebar";
 import { ContourList } from "./ContourList";
 
@@ -8,6 +8,7 @@ interface SidebarProps {
     tool: string;
     contours: any;
     selectContour: (index:number) => void;
+    setTool: (tool:string) => void;
     selectedContour: number;
 }
 
@@ -18,11 +19,12 @@ export class Sidebar extends React.Component<SidebarProps,{}> {
             <div id="sidebar">
                 <CreateToolSidebar isActive={this.props.tool == "create"}/>
                 <SelectToolSidebar isActive={this.props.tool == "select"}/>
-                <MoveToolSidebar isActive={this.props.tool == "move"}/>
+                <EditToolSidebar isActive={this.props.tool == "edit"}/>
                 <ContourList
                     contours={this.props.contours}
                     selectContour={this.props.selectContour}
                     selectedContour={this.props.selectedContour}
+                    setTool={this.props.setTool}
                     />
             </div>
         );